breaking: drop deprecated devstore
